webapp/requirements.txt
Django==4.1.5
git+https://github.com/azure-samples/ms-identity-python-utilities@main
django-storages[azure]==1.13.2
pymongo==3.12.3
djongo==1.3.6
django-cors-headers==3.13.0
pytz==2022.7
Pillow==9.4.0
selenium==4.7.2
opencensus-ext-django==0.8.0
opencensus-ext-azure==1.1.7